requirements.txt
alembic==1.7.5
aniso8601==9.0.1
apispec==6.0.2
apispec-webframeworks==0.5.2
APScheduler==3.9.1
attrs==22.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
beautifulsoup4==4.11.1
binary==1.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
docopt==0.6.2
ffmpeg==1.4
ffmpeg-python==0.2.0
flasgger==0.9.5
Flask==2.1.1
flask-apispec==0.11.4
Flask-APScheduler==1.12.4
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==4.11.1
fonttools==4.28.5
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Js2Py==0.71
jsonschema==4.17.3
kiwisolver==1.3.2
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.0
matplotlib==3.5.1
mistune==2.0.5
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
pipwin==0.5.2
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.3
pycparser==2.21
pydub==0.25.1
pyjsparser==2.7.1
PyJWT==2.4.0
pyparsing==3.0.6
PyPrind==2.11.3
pyrsistent==0.19.3
pySmartDL==1.3.4
PySoundFile==0.9.0.post1
pyTelegramBotAPI==4.2.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.27
tqdm==4.64.1
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.7
wavio==0.0.4
webargs==8.2.0
Werkzeug==2.0.3
zipp==3.6.0
